# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Gym Manager API specification
servers:
  - url: /api
    description: Base API path
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  # ---------------------------------------------------------------------------
  # Reusable schemas
  # ---------------------------------------------------------------------------
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    StaffMember:
      type: object
      properties:
        id:
          type: string
          format: uuid
        gymId:
          type: string
        name:
          type: string
        username:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [id, gymId, name, username, createdAt]
    StaffLoginRequest:
      type: object
      properties:
        gymId:
          type: string
          minLength: 1
        username:
          type: string
          minLength: 1
        pin:
          type: string
          minLength: 1
      required: [gymId, username, pin]
    StaffAuthResponse:
      type: object
      properties:
        token:
          type: string
        staff:
          type: object
          properties:
            id:
              type: string
              format: uuid
            gymId:
              type: string
            name:
              type: string
            role:
              type: string
              enum: [staff]
          required: [id, gymId, name, role]
      required: [token, staff]
    CreateStaffRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
        username:
          type: string
          minLength: 2
          maxLength: 30
          pattern: "^[a-z0-9_]+$"
          description: Lowercase letters, numbers, and underscores only
        pin:
          type: string
          pattern: "^\\d{4,6}$"
          description: 4–6 numeric digits
      required: [name, username, pin]
    StaffListResponse:
      type: object
      properties:
        staff:
          type: array
          items:
            $ref: "#/components/schemas/StaffMember"
      required: [staff]

# ---------------------------------------------------------------------------
# Paths
# ---------------------------------------------------------------------------
paths:
...
  /staff/login:
    post:
      operationId: loginStaff
      tags: [staff]
      summary: Log in as a staff member
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StaffLoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StaffAuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /staff:
    get:
      operationId: listStaff
      tags: [staff]
      summary: List all staff for the authenticated gym (owner only)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of staff members
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StaffListResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Not authorised (owner role required)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    post:
      operationId: createStaff
      tags: [staff]
      summary: Add a new staff member (owner only)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateStaffRequest"
      responses:
        "201":
          description: Staff member created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StaffMember"
        "400":
          description: Validation failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Not authorised (owner role required)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Username already taken in this gym
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /staff/{staffId}:
    parameters:
      - name: staffId
        in: path
        required: true
        schema:
          type: string
          format: uuid

    delete:
      operationId: deleteStaff
      tags: [staff]
      summary: Remove a staff member (owner only)
      security:
        - bearerAuth: []
      responses:
        "204":
          description: Staff member deleted
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Not authorised (owner role required)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Staff member not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"