# lib/api-zod/src/generated/types/staffAuthResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Gym Manager API specification
 * OpenAPI spec version: 0.1.0
 */
import type { StaffAuthResponseStaff } from "./staffAuthResponseStaff";

export interface StaffAuthResponse {
  token: string;
  staff: StaffAuthResponseStaff;
}