# lib/api-zod/src/generated/types/staffLoginRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Gym Manager API specification
 * OpenAPI spec version: 0.1.0
 */

export interface StaffLoginRequest {
  /** @minLength 1 */
  gymId: string;
  /** @minLength 1 */
  username: string;
  /** @minLength 1 */
  pin: string;
}